refactor: Reorganize user login and user management
